Correct type

+++ module/spl/spl-processor.c
@@ -36,7 +36,7 @@ getcpuid()
 	return ((uint32_t)cpu_number());
 }
 
-unsigned long long spl_cpuid_features(void)
+uint64_t spl_cpuid_features(void)
 {
 	i386_cpu_info_t *info;
 
@@ -44,7 +44,7 @@ unsigned long long spl_cpuid_features(void)
 	return info->cpuid_features;
 }
 
-unsigned long long spl_cpuid_leaf7_features(void)
+uint64_t spl_cpuid_leaf7_features(void)
 {
 	i386_cpu_info_t *info;
